Changed mallocMagic() argument to size_t.  However, this is probably not useful so long as Tcl_Alloc() has (unsigned int) for an argument. The more important investigation is probably to determine if there is a way to keep csa2_list from growing to absurdly large sizes on connectivity checks.

+++ utils/malloc.c
@@ -92,7 +92,7 @@ static char *freeDelayedItem = NULL;
 
 void *
 mallocMagic(nbytes)
-    unsigned int nbytes;
+    size_t nbytes;
 {
     void *p;
     
@@ -150,12 +150,12 @@ freeMagic(cp)
 
 void *
 callocMagic(nbytes)
-    unsigned int nbytes;
+    size_t nbytes;
 {
     void *cp;
 
     cp = mallocMagic(nbytes);
-    bzero(cp, (int) nbytes);
+    bzero(cp, (size_t) nbytes);
 
     return (cp);
 }
